requirements.txt
python-binance>=1.0.19
websockets>=11.0.3
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.18.0
